+++ src/tilda/types/Type_CharPrimitiveNull.java
@@ -19,7 +19,7 @@ package tilda.types;
 
 public class Type_CharPrimitiveNull extends Type_CharPrimitive implements Nullable
   {
-    public Type_CharPrimitiveNull(String SchemaName, String TableName, String ColumnName, long Count, String Description)
+    public Type_CharPrimitiveNull(String SchemaName, String TableName, String ColumnName, int Count, String Description)
       {
         super(SchemaName, TableName, ColumnName, Count, Description);
       }
