+++ src/tilda/types/Type_JsonCollection.java
@@ -20,7 +20,7 @@ import tilda.enums.ColumnType;
 
 public class Type_JsonCollection extends ColumnDefinition
   {
-    public Type_JsonCollection(String SchemaName, String TableName, String ColumnName, long Count, String Description)
+    public Type_JsonCollection(String SchemaName, String TableName, String ColumnName, int Count, String Description)
       {
         super(SchemaName, TableName, ColumnName, Count, ColumnType.JSON, true, Description);
       }
