+++ src/tilda/types/Type_LongCollection.java
@@ -20,7 +20,7 @@ import tilda.enums.ColumnType;
 
 public class Type_LongCollection extends ColumnDefinition
   {
-    public Type_LongCollection(String SchemaName, String TableName, String ColumnName, long Count, String Description)
+    public Type_LongCollection(String SchemaName, String TableName, String ColumnName, int Count, String Description)
       {
         super(SchemaName, TableName, ColumnName, Count, ColumnType.LONG, true, Description);
       }
